+++ src/types.ts
@@ -1,0 +1,1 @@
+export type Resolution = '240p' | '360p' | '480p' | '720p' | '1080p';
